requirements.txt
fastapi
uvicorn
sqlalchemy
python-dotenv
requests
passlib[bcrypt]
python-jose
redis>=4.2
python-multipart
pydantic[email]
pytest
httpx